API developers now only use the package `xcube.server.api`
